avformat/http: Add max_request_size option
In some cases, webservers may return content more quickly if one
sends many small requests rather than one big request. Clients may
wish to opt-in to this workaround.

For this purpose, add a max_request_size AVOption, which returns
EOF if max_request_size is exceeded, causing a new request with
an offset to be issued.

In particular, this hack is useful for mpv, where it can be used to
bypass YouTube's throttling.

Signed-off-by: Nicolas Frattaroli <ffmpeg@fratti.ch>
Signed-off-by: tuilakhanh <bacdau852@gmail.com>